Move table of contents logic out of App
